pyproject.toml
[project]
name = "codeask"
version = "0.1.0"
description = "API that answers product questions by exploring a GitHub repo's codebase"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "anthropic>=0.49",
    "pydantic-settings>=2.7",
    "gitpython>=3.1",
    "slack-bolt>=1.20",
]
